+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added action `enqueueLinksByClicking`.
+- Added previously deleted option `baseUrl` for the action `enqueueLinks`
+
+### Changed
+- Updated dependency `crawlee` to the version `^3.4.0`.
+
+### Fixed
+- Fixed cleanup after failed crawling process.
+- Action `clickWithRedirect` changes `request.loadedUrl` to the current URL after a redirect.
 
 ## 0.2.1 - 2023-06-14
 ### Added